requirements.txt
pandas==0.24.2
numpy==1.16.4
biopython==1.73
